Combine processRecentEpisodes functions
